wip - convert transformation into operation.
