+++ src/Contract/Collection.php
@@ -6,6 +6,7 @@ namespace loophp\collection\Contract;
 
 use IteratorAggregate;
 use JsonSerializable;
+use loophp\collection\Contract\Operation\Allable;
 use loophp\collection\Contract\Operation\Appendable;
 use loophp\collection\Contract\Operation\Applyable;
 use loophp\collection\Contract\Operation\Associateable;
@@ -16,19 +17,26 @@ use loophp\collection\Contract\Operation\Columnable;
 use loophp\collection\Contract\Operation\Combinateable;
 use loophp\collection\Contract\Operation\Combineable;
 use loophp\collection\Contract\Operation\Compactable;
+use loophp\collection\Contract\Operation\Containsable;
 use loophp\collection\Contract\Operation\Cycleable;
 use loophp\collection\Contract\Operation\Diffable;
 use loophp\collection\Contract\Operation\Diffkeysable;
 use loophp\collection\Contract\Operation\Distinctable;
 use loophp\collection\Contract\Operation\Explodeable;
+use loophp\collection\Contract\Operation\Falsyable;
 use loophp\collection\Contract\Operation\Filterable;
 use loophp\collection\Contract\Operation\Firstable;
 use loophp\collection\Contract\Operation\Flattenable;
 use loophp\collection\Contract\Operation\Flipable;
+use loophp\collection\Contract\Operation\FoldLeftable;
+use loophp\collection\Contract\Operation\FoldRightable;
 use loophp\collection\Contract\Operation\Forgetable;
 use loophp\collection\Contract\Operation\Frequencyable;
+use loophp\collection\Contract\Operation\Getable;
 use loophp\collection\Contract\Operation\Groupable;
 use loophp\collection\Contract\Operation\Headable;
+use loophp\collection\Contract\Operation\Hasable;
+use loophp\collection\Contract\Operation\Implodeable;
 use loophp\collection\Contract\Operation\Intersectable;
 use loophp\collection\Contract\Operation\Intersectkeysable;
 use loophp\collection\Contract\Operation\Intersperseable;
@@ -41,6 +49,7 @@ use loophp\collection\Contract\Operation\Mapable;
 use loophp\collection\Contract\Operation\Mergeable;
 use loophp\collection\Contract\Operation\Normalizeable;
 use loophp\collection\Contract\Operation\Nthable;
+use loophp\collection\Contract\Operation\Nullsyable;
 use loophp\collection\Contract\Operation\Onlyable;
 use loophp\collection\Contract\Operation\Padable;
 use loophp\collection\Contract\Operation\Pairable;
@@ -50,9 +59,11 @@ use loophp\collection\Contract\Operation\Prependable;
 use loophp\collection\Contract\Operation\Productable;
 use loophp\collection\Contract\Operation\Randomable;
 use loophp\collection\Contract\Operation\Rangeable;
+use loophp\collection\Contract\Operation\Reduceable;
 use loophp\collection\Contract\Operation\Reductionable;
 use loophp\collection\Contract\Operation\Reverseable;
 use loophp\collection\Contract\Operation\RSampleable;
+use loophp\collection\Contract\Operation\Runable;
 use loophp\collection\Contract\Operation\Scaleable;
 use loophp\collection\Contract\Operation\Shuffleable;
 use loophp\collection\Contract\Operation\Sinceable;
@@ -62,26 +73,15 @@ use loophp\collection\Contract\Operation\Sortable;
 use loophp\collection\Contract\Operation\Splitable;
 use loophp\collection\Contract\Operation\Tailable;
 use loophp\collection\Contract\Operation\Timesable;
+use loophp\collection\Contract\Operation\Transformable;
 use loophp\collection\Contract\Operation\Transposeable;
+use loophp\collection\Contract\Operation\Truthyable;
 use loophp\collection\Contract\Operation\Unpairable;
 use loophp\collection\Contract\Operation\Untilable;
 use loophp\collection\Contract\Operation\Unwrapable;
 use loophp\collection\Contract\Operation\Windowable;
 use loophp\collection\Contract\Operation\Wrapable;
 use loophp\collection\Contract\Operation\Zipable;
-use loophp\collection\Contract\Transformation\Allable;
-use loophp\collection\Contract\Transformation\Containsable;
-use loophp\collection\Contract\Transformation\Falsyable;
-use loophp\collection\Contract\Transformation\FoldLeftable;
-use loophp\collection\Contract\Transformation\FoldRightable;
-use loophp\collection\Contract\Transformation\Getable;
-use loophp\collection\Contract\Transformation\Hasable;
-use loophp\collection\Contract\Transformation\Implodeable;
-use loophp\collection\Contract\Transformation\Nullsyable;
-use loophp\collection\Contract\Transformation\Reduceable;
-use loophp\collection\Contract\Transformation\Runable;
-use loophp\collection\Contract\Transformation\Transformable;
-use loophp\collection\Contract\Transformation\Truthyable;
 
 /**
  * @psalm-template TKey
