+++ src/Contract/Operation/FoldRightable.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace loophp\collection\Contract\Transformation;
+namespace loophp\collection\Contract\Operation;
 
 /**
  * @psalm-template TKey
  * @psalm-template TKey of array-key
  * @psalm-template T
  */
-interface FoldLeftable
+interface FoldRightable
 {
     /**
-     * Fold the collection from the left to the right.
+     * Fold the collection from the right to the left.
      *
      * @param mixed $initial
      * @psalm-param T|null $initial
@@ -20,5 +20,5 @@ interface FoldLeftable
      * @return mixed
      * @psalm-return T|null
      */
-    public function foldLeft(callable $callback, $initial = null);
+    public function foldRight(callable $callback, $initial = null);
 }
