+++ src/Operation/Contains.php
@@ -2,20 +2,21 @@
 
 declare(strict_types=1);
 
-namespace loophp\collection\Transformation;
+namespace loophp\collection\Operation;
 
 use ArrayIterator;
+use Closure;
 use Iterator;
-use loophp\collection\Contract\Transformation;
+use loophp\collection\Contract\Operation;
 
 /**
  * @psalm-template TKey
  * @psalm-template TKey of array-key
  * @psalm-template T
  *
- * @implements Transformation<TKey, T>
+ * @implements Operation<TKey, T>
  */
-final class Contains extends AbstractTransformation implements Transformation
+final class Contains extends AbstractOperation implements Operation
 {
     /**
      * @param mixed ...$values
@@ -26,7 +27,7 @@ final class Contains extends AbstractTransformation implements Transformation
         $this->storage['values'] = new ArrayIterator($values);
     }
 
-    public function __invoke()
+    public function __invoke(): Closure
     {
         return static function (Iterator $collection, ArrayIterator $values): bool {
             return (new Transform(new FoldLeft(
