+++ src/Operation/Flatten.php
@@ -9,7 +9,6 @@ use Closure;
 use Generator;
 use Iterator;
 use loophp\collection\Contract\Operation;
-use loophp\collection\Transformation\Run;
 
 use function is_array;
