+++ src/Operation/Sort.php
@@ -10,7 +10,6 @@ use Exception;
 use Generator;
 use Iterator;
 use loophp\collection\Contract\Operation;
-use loophp\collection\Transformation\Run;
 
 /**
  * @psalm-template TKey
