+++ src/Operation/Transform.php
@@ -2,24 +2,25 @@
 
 declare(strict_types=1);
 
-namespace loophp\collection\Transformation;
+namespace loophp\collection\Operation;
 
+use Closure;
 use Iterator;
-use loophp\collection\Contract\Transformation;
+use loophp\collection\Contract\Operation;
 
 /**
  * @psalm-template TKey
  * @psalm-template TKey of array-key
  * @psalm-template T
  *
- * @implements Transformation<TKey, T>
+ * @implements Operation<TKey, T>
  */
-final class Transform extends AbstractTransformation implements Transformation
+final class Transform extends AbstractOperation implements Operation
 {
     /**
-     * @param \loophp\collection\Contract\Transformation<TKey, T> ...$transformers
+     * @param \loophp\collection\Contract\Operation<TKey, T> ...$transformers
      */
-    public function __construct(Transformation ...$transformers)
+    public function __construct(Operation ...$transformers)
     {
         $this->storage['transformers'] = $transformers;
         $this->storage['wrapper'] = static function (callable $callable, ...$arguments) {
@@ -27,16 +28,16 @@ final class Transform extends AbstractTransformation implements Transformation
         };
     }
 
-    public function __invoke()
+    public function __invoke(): Closure
     {
         return function (Iterator $collection) {
             return array_reduce(
                 $this->get('transformers', []),
-                function (Iterator $collection, Transformation $transformer) {
+                function (Iterator $collection, Operation $operation) {
                     return ($this->get('wrapper'))(
-                        $transformer(),
+                        $operation(),
                         $collection,
-                        ...array_values($transformer->getArguments())
+                        ...array_values($operation->getArguments())
                     );
                 },
                 $collection
