+++ src/Operation/Truthy.php
@@ -2,25 +2,26 @@
 
 declare(strict_types=1);
 
-namespace loophp\collection\Transformation;
+namespace loophp\collection\Operation;
 
+use Closure;
 use Iterator;
-use loophp\collection\Contract\Transformation;
+use loophp\collection\Contract\Operation;
 
 /**
  * @psalm-template TKey
  * @psalm-template TKey of array-key
  * @psalm-template T
  *
- * @implements Transformation<TKey, T>
+ * @implements Operation<TKey, T>
  */
-final class Falsy extends AbstractTransformation implements Transformation
+final class Truthy extends AbstractOperation implements Operation
 {
-    public function __invoke()
+    public function __invoke(): Closure
     {
         return static function (Iterator $collection): bool {
             foreach ($collection as $key => $value) {
-                if (false !== (bool) $value) {
+                if (false === (bool) $value) {
                     return false;
                 }
             }
